refactor(MovieHeader): Header 내 Logo를 컴포넌트로 분리
